implement check features before creating file and book sale

+++ backend/src/domain/usecase/profile/delete-user-with-profile.ts
@@ -1,4 +1,6 @@
 import { UserProfile, UserProfileCreate } from "../../entity/user-profile";
+import { CustomError } from "../../error/custom-error";
+import { ErrorType } from "../../error/error-type";
 import { IUserProfileRepository } from "../../repository/iuser-profile-repository";
 import { IUserRepository } from "../../repository/iuser-repository";
 
@@ -8,8 +10,11 @@ export class DeleteUser {
     private readonly userRepository: IUserRepository
   ) {}
 
-  async execute(uid: string): Promise<UserProfile | null> {
+  async execute(uid: string): Promise<UserProfile> {
     const user = await this.userProfileRepository.getUserProfile(uid);
+    if (user == null) {
+      throw new CustomError("You don't have any profile.", ErrorType.NOT_FOUND);
+    }
     await this.userRepository.deleteUser(uid);
     return user;
   }
